requirements.txt
requests
python-dotenv
slugify
customtkinter
pillow
ddgs
youtube-transcript-api
google-genai
beautifulsoup4
